Release 0.0.6 (#92)
* Update flyway to v10.21.0 (#50)


* Update dependency org.junit.jupiter:junit-jupiter-api to v5.11.3 (#52)


* Update actions/setup-java action to v4 (#74)


* Update jackson.version to v2.18.1 (#73)


* Update jackson monorepo to v2.18.1 (#58)


* Update spring boot to v3.3.5 (#57)


* Update dependency org.springframework:spring-core to v6.2.0 (#56)


* Update dependency org.jetbrains:annotations to v26 (#54)


* 43 max failed sing in requests (#81)

* Add password migration, max login attempts and recover password url expiration

* Remove migration leftovers

* Remove form and basic login

* Add separate delete code url

* Update html views

* 75 add profile self description (#87)

* Add user profile and username change

* Update refresh token to use userId

* Fix username change nullpointer

* Add deleting pin objects from minio when deleting group or user

* Make exception runtime level to be caught in exceptionHandler

* Use MonaApi release version

* Fix api migration

* 82 add like per user endpoint and like points (#90)

* Add like caching

* Fix count by creation user queries

* Update api version to release

---------

+++ src/main/kotlin/de/lrprojects/monaserver/service/impl/UserServiceImpl.kt
@@ -1,11 +1,14 @@
 package de.lrprojects.monaserver.service.impl
 
 import de.lrprojects.monaserver.entity.User
+import de.lrprojects.monaserver.excepetion.TimeExpiredException
+import de.lrprojects.monaserver.excepetion.UserExistsException
 import de.lrprojects.monaserver.excepetion.UserNotFoundException
 import de.lrprojects.monaserver.helper.ImageHelper
 import de.lrprojects.monaserver.repository.UserRepository
 import de.lrprojects.monaserver.security.TokenHelper
 import de.lrprojects.monaserver.service.api.ObjectService
+import de.lrprojects.monaserver.service.api.PinService
 import de.lrprojects.monaserver.service.api.RefreshTokenService
 import de.lrprojects.monaserver.service.api.UserService
 import de.lrprojects.monaserver.service.impl.ObjectServiceImpl.Companion.getUserFileProfile
@@ -17,17 +20,21 @@ import jakarta.persistence.EntityNotFoundException
 import org.springframework.cache.annotation.CacheEvict
 import org.springframework.cache.annotation.Cacheable
 import org.springframework.cache.annotation.Caching
+import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import java.time.OffsetDateTime
 import java.util.*
 
 @Service
 class UserServiceImpl(
-    val userRepository: UserRepository,
-    val refreshTokenService: RefreshTokenService,
-    val imageHelper: ImageHelper,
-    val tokenHelper: TokenHelper,
-    private val objectService: ObjectService
+    private val userRepository: UserRepository,
+    private val refreshTokenService: RefreshTokenService,
+    private val imageHelper: ImageHelper,
+    private val tokenHelper: TokenHelper,
+    private val objectService: ObjectService,
+    private val passwordEncoder: PasswordEncoder,
+    private val pinService: PinService
 ): UserService {
 
     @Transactional
@@ -46,8 +53,13 @@ class UserServiceImpl(
     override fun deleteUser(userId: UUID, code: Int) {
         val user = userRepository.findByIdAndCode(userId, code.toString())
             .orElseThrow { EntityNotFoundException("user and code in this combination do not exist") }
+        if (OffsetDateTime.now().isAfter(user.codeExpiration!!)) {
+            throw TimeExpiredException("code is expired")
+        }
+        val ids = pinService.getUserPins(user)
         refreshTokenService.invalidateTokens(user)
         userRepository.delete(user)
+        pinService.deleteObjectsByList(ids)
     }
 
     override fun getUserProfileImage(userId: UUID): String? {
@@ -73,20 +85,33 @@ class UserServiceImpl(
     }
 
     @CacheEvict(value = ["users"], key = "#userId")
+    @Transactional
     override fun updateUser(userId: UUID, user: UserUpdateDto): TokenResponseDto? {
         val userEntity =  getUser(userId)
         var responseDto: TokenResponseDto? = null
         if (user.email != null) {
             userEntity.email = user.email
             userEntity.code = null
+            userEntity.codeExpiration = null
         }
         if (user.password != null) {
-            userEntity.password = user.password
+            userEntity.password = passwordEncoder.encode(user.password)
             userEntity.resetPasswordUrl = null
+            userEntity.resetPasswordExpiration = null
+            userEntity.failedLoginAttempts = 0
             refreshTokenService.invalidateTokens(userEntity)
-            val accessToken = tokenHelper.generateToken(userEntity.username)
+            val accessToken = tokenHelper.generateToken(userEntity.id!!)
             val refreshToken = refreshTokenService.createRefreshToken(userEntity)
             responseDto = TokenResponseDto(refreshToken.token, accessToken, userEntity.id!!)
+        }
+        if (user.description != null) {
+            userEntity.description = user.description
+        }
+        if (user.username != null) {
+            if (userEntity.lastUsernameUpdate != null && OffsetDateTime.now().isBefore(userEntity.lastUsernameUpdate!!.plusDays(USERNAME_CHANGE_TIMEOUT))) { throw TimeExpiredException("username can only be changed once every 14 days") }
+            userRepository.findByUsername(user.username).ifPresent { throw UserExistsException("user with this username already exists") }
+            userEntity.username = user.username
+            userEntity.lastUsernameUpdate = OffsetDateTime.now()
         }
         userRepository.save(userEntity)
         return responseDto
@@ -115,7 +140,26 @@ class UserServiceImpl(
         if (list == null) {
             throw UserNotFoundException("user with this reset url does not exist")
         } else {
-            return list
+            if (OffsetDateTime.now().isBefore(list.resetPasswordExpiration!!)) {
+                return list
+            }
+            throw TimeExpiredException("reset url is expired")
         }
+    }
+
+    override fun getUserByDeletionUrl(deletionUrl: String): User {
+        val list = userRepository.findByDeletionUrl(deletionUrl).firstOrNull();
+        if (list == null) {
+            throw UserNotFoundException("user with this reset url does not exist")
+        } else {
+            if (OffsetDateTime.now().isBefore(list.codeExpiration!!)) {
+                return list
+            }
+            throw TimeExpiredException("deletion url is expired")
+        }
+    }
+
+    companion object {
+        private const val USERNAME_CHANGE_TIMEOUT: Long = 14
     }
 }
